backend/requirements.txt
# ── Core (required) ──────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
websockets>=12.0
aiohttp>=3.9.0
aiosmtplib>=3.0.0
jinja2>=3.1.2
fpdf2>=2.7.0
httpx>=0.25.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
google-genai>=1.0.0

# ── AI / ML (install separately if needed) ──
# pip install sentence-transformers openai-whisper opencv-python-headless deepface
sentence-transformers>=2.2.2
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ── New AI Components ──────────────────────
PyPDF2>=3.0.0
python-docx>=1.0.0
networkx>=3.2.0
shap>=0.43.0
gymnasium>=0.29.0
deepface>=0.0.80
opencv-python-headless>=4.8.0
